requirements.txt
clean-fid
torch-ema
ballpark~=1.4.0
argparse~=1.4.0
pyyaml
typing~=3.6.4
omegaconf~=2.2
pillow~=8.4.0
wandb~=0.12.2
hydra~=2.5
pytorch-lightning==1.7.7
tqdm~=4.57.0
torchmetrics~=0.7.0
pyrender
scipy
setuptools~=49.6.0
numpy>=1.20
matplotlib~=3.7.1
opencv-python~=3.4.9.33
scikit-image~=0.19.1
ninja
imageio
imageio-ffmpeg
hydra-core==1.2.0
torchmetrics[image]
urllib3<2.0.0
pandas~=2.0.2
chumpy~=0.70
